+++ website/src/docs/other_examples.js
@@ -163,6 +163,29 @@ export default function Example() {
     jsx: <DatePicker monthYearSeparator="|" {...otherProps} />,
   };
 
+  const formatMonthAndYear = {
+    title: "Custom Month & Year In Header",
+    code: `${localeImport}<DatePicker 
+  formatMonth={(month, year) => {
+    return "Month " + month;
+  }}
+  formatYear={(year, month) => {
+    return "Year " + year;
+  }}
+/>`,
+    jsx: (
+      <DatePicker
+        formatMonth={(month, year) => {
+          return "Month " + month;
+        }}
+        formatYear={(year, month) => {
+          return "Year " + year;
+        }}
+        {...otherProps}
+      />
+    ),
+  };
+
   const weekNumbers = {
     title: "Displaying Week Numbers",
     code: `${localeImport}<DatePicker 
@@ -347,6 +370,7 @@ export default function Example() {
     weekStartDayIndex,
     otherDays,
     monthYearSeparator,
+    formatMonthAndYear,
     weekNumbers,
     scroll,
     fullYear,